rename commands accordingly
