+++ src/main/java/frc/robot/autos/RedAlliance/Red2AmpAuto.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.autos.BlueAlliance;
+package frc.robot.autos.RedAlliance;
 
 
 import edu.wpi.first.wpilibj2.command.InstantCommand;
@@ -23,19 +23,22 @@ import frc.robot.trajectories.AmpTrajectories;
 // NOTE:  Consider using this command inline, rather than writing a subclass.  For more
 // information, see:
 // https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
-public class Blue3AmpAuto extends SequentialCommandGroup {
+public class Red2AmpAuto extends SequentialCommandGroup {
         
   SwerveSubsystem swerveSubsystem = SwerveSubsystem.getInstance();
   // contruct command to follow trajectory
-      /*
-       * 
-       * BLUE Autos
-       * 
-       */
+  SwerveControllerCommand orginToAmp = new SwerveControllerCommand(
+    AmpTrajectories.tragOriginToAmp, 
+    swerveSubsystem::getPose, // Coords
+    DriveConstants.kDriveKinematics, 
+    AutoConstants.xController, 
+    AutoConstants.yController,
+    AutoConstants.thetaController,
+    swerveSubsystem::setModuleStates, // Function to translate speeds to the modules
+    swerveSubsystem);
 
-  // contruct command to follow trajectory
-  SwerveControllerCommand blueOrginToAmp = new SwerveControllerCommand(
-    AmpTrajectories.tragBlueOriginToAmp, 
+  SwerveControllerCommand ampToSpeaker = new SwerveControllerCommand(
+    AmpTrajectories.tragAmpToSpeakerNote, 
     swerveSubsystem::getPose, // Coords
     DriveConstants.kDriveKinematics, 
     AutoConstants.xController, 
@@ -45,9 +48,8 @@ public class Blue3AmpAuto extends SequentialCommandGroup {
     swerveSubsystem);
 
 
-
-  SwerveControllerCommand blueAmpToAmpN = new SwerveControllerCommand(
-    AmpTrajectories.tragBlueAmpToAmpNote, 
+  SwerveControllerCommand ampToAmpN = new SwerveControllerCommand(
+    AmpTrajectories.tragAmpToAmpNote, 
     swerveSubsystem::getPose, // Coords
     DriveConstants.kDriveKinematics, 
     AutoConstants.xController, 
@@ -56,8 +58,8 @@ public class Blue3AmpAuto extends SequentialCommandGroup {
     swerveSubsystem::setModuleStates, // Function to translate speeds to the modules
     swerveSubsystem);
 
-  SwerveControllerCommand blueAmpNToAmp = new SwerveControllerCommand(
-    AmpTrajectories.tragBlueAmpNoteToAmp, 
+  SwerveControllerCommand ampNToAmp = new SwerveControllerCommand(
+    AmpTrajectories.tragAmpNoteToAmp, 
     swerveSubsystem::getPose, // Coords
     DriveConstants.kDriveKinematics, 
     AutoConstants.xController, 
@@ -66,33 +68,28 @@ public class Blue3AmpAuto extends SequentialCommandGroup {
     swerveSubsystem::setModuleStates, // Function to translate speeds to the modules
     swerveSubsystem);
   
-  SwerveControllerCommand blueSpeakerNoteToAmp = new SwerveControllerCommand(
-      AmpTrajectories.tragBlueSpeakerNoteToAmp, 
-      swerveSubsystem::getPose, // Coords
-      DriveConstants.kDriveKinematics, 
-      AutoConstants.xController, 
-      AutoConstants.yController,
-      AutoConstants.thetaController,
-      swerveSubsystem::setModuleStates, // Function to translate speeds to the modules
-      swerveSubsystem);
-  
-      SwerveControllerCommand blueAmpToSpeaker = new SwerveControllerCommand(
-        AmpTrajectories.tragBlueAmpToSpeakerNote, 
-        swerveSubsystem::getPose, // Coords
-        DriveConstants.kDriveKinematics, 
-        AutoConstants.xController, 
-        AutoConstants.yController,
-        AutoConstants.thetaController,
-        swerveSubsystem::setModuleStates, // Function to translate speeds to the modules
-        swerveSubsystem);
+  SwerveControllerCommand speakerNoteToAmp = new SwerveControllerCommand(
+    AmpTrajectories.tragSpeakerNoteToAmp, 
+    swerveSubsystem::getPose, // Coords
+    DriveConstants.kDriveKinematics, 
+    AutoConstants.xController, 
+    AutoConstants.yController,
+    AutoConstants.thetaController,
+    swerveSubsystem::setModuleStates, // Function to translate speeds to the modules
+    swerveSubsystem);
 
+  /** Creates a new Red3AmpAuto. */
+ 
+  /*
+   * AMP Trajectories
+   */
 
-  public Blue3AmpAuto() {
+  public Red2AmpAuto() {
     addCommands(
       new InstantCommand(() -> swerveSubsystem.resetOdometry(AmpTrajectories.tragOriginToAmp.getInitialPose())),
       // Go to AMP while pivoting to AMP Pos.
       new ParallelCommandGroup(
-        blueOrginToAmp, 
+        orginToAmp, 
         new PIDPivotToPosition(Constants.PivoterConstants.kAmpPivoterRotations)
       ),
       new InstantCommand(() -> swerveSubsystem.stopModules()),
@@ -114,7 +111,7 @@ public class Blue3AmpAuto extends SequentialCommandGroup {
           new IntakeFromGround(),
           new WaitCommand(2)
         ),
-        blueAmpToAmpN
+        ampToAmpN
       ),
       new InstantCommand(() -> swerveSubsystem.stopModules()),
 
@@ -124,7 +121,7 @@ public class Blue3AmpAuto extends SequentialCommandGroup {
       new WaitCommand(0.25),
       new InstantCommand(() -> swerveSubsystem.resetOdometry(AmpTrajectories.tragAmpNoteToAmp.getInitialPose())),
       new ParallelCommandGroup(
-        blueAmpNToAmp, 
+        ampNToAmp, 
         new PIDPivotToPosition(Constants.PivoterConstants.kAmpPivoterRotations)
       ),
       new InstantCommand(() -> swerveSubsystem.stopModules()),
@@ -138,34 +135,25 @@ public class Blue3AmpAuto extends SequentialCommandGroup {
     
       // Move arm down to rest pos
       new PIDPivotToPosition(0)
-      
+
       );
-    /*
-     * 
-     * ORIGINAL
-     * 
-     */
-    // addCommands(
-    //   // Reset odometry to starting pose. 
-    //   new InstantCommand(() -> swerveSubsystem.resetOdometry(AmpTrajectories.tragBlueOriginToAmp.getInitialPose())),
-    //   blueOrginToAmp,
-    //   new InstantCommand(() -> swerveSubsystem.stopModules()),
-    //   new WaitCommand(0.25),
-    //   new InstantCommand(() -> swerveSubsystem.resetOdometry(AmpTrajectories.tragBlueAmpToAmpNote.getInitialPose())),
-    //   blueAmpToAmpN,
-    //   new InstantCommand(() -> swerveSubsystem.stopModules()),
-    //   new WaitCommand(0.25),
-    //   new InstantCommand(() -> swerveSubsystem.resetOdometry(AmpTrajectories.tragBlueAmpNoteToAmp.getInitialPose())),
-    //   blueAmpNToAmp,
-    //   new InstantCommand(() -> swerveSubsystem.stopModules()),
-    //   new WaitCommand(0.25),
-    //   new InstantCommand(() -> swerveSubsystem.resetOdometry(AmpTrajectories.tragBlueAmpToSpeakerNote.getInitialPose())),
-    //   blueAmpToSpeaker,
-    //   new InstantCommand(() -> swerveSubsystem.stopModules()),
-    //   new WaitCommand(0.25),
-    //   new InstantCommand(() -> swerveSubsystem.resetOdometry(AmpTrajectories.tragBlueSpeakerNoteToAmp.getInitialPose())),
-    //   blueSpeakerNoteToAmp,
-    //   new InstantCommand(() -> swerveSubsystem.stopModules())
-    //   );
+
+      /*
+       * 
+       * ORIGINAL
+       * 
+       */
+
+      // new WaitCommand(0.25),
+      // new InstantCommand(() -> swerveSubsystem.resetOdometry(AmpTrajectories.tragAmpNoteToAmp.getInitialPose())),
+      // ampNToAmp,
+      // new InstantCommand(() -> swerveSubsystem.stopModules()),
+      // new WaitCommand(0.25),
+      // new InstantCommand(() -> swerveSubsystem.resetOdometry(AmpTrajectories.tragAmpToSpeakerNote.getInitialPose())),
+      // ampToSpeaker,
+      // new InstantCommand(() -> swerveSubsystem.stopModules()),
+      // new WaitCommand(0.25),
+      // new InstantCommand(() -> swerveSubsystem.resetOdometry(AmpTrajectories.tragSpeakerNoteToAmp.getInitialPose())),
+      // speakerNoteToAmp,
   }
 }
